requirements.txt
# Core LangChain/LangGraph packages (v1.0)
# IMPORTANT: v1.0 requires Python 3.10+
langchain>=1.0.0
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-openai>=1.0.0
langchain-anthropic>=1.0.0

# Vision and Image Processing
opencv-python==4.10.0.84
Pillow==10.4.0
numpy==1.26.4

# OCR (Optical Character Recognition)
easyocr==1.7.1
pytesseract==0.3.13

# Face Detection (will use later)
# mediapipe==0.10.14  # Uncomment when needed

# Environment and Configuration
python-dotenv==1.0.1

# Utilities
pydantic==2.9.2
pydantic-settings==2.5.2

# Optional: For better terminal output
rich==13.9.2
colorama==0.4.6

# Development tools (optional)
pytest==8.3.3
black==24.10.0